Bugfix: No new Item will be created
Problem:
When you want to create a new entry by selecting an
existing entry before you click on "create new entry"
the selected item will be updated instead of creating a
new entry.

Description:
This bug was because the current logic distinguish
between creating a new or update an existing entry
by setting a reference to an existing item. Therefore
the program thought you want to update the existing item.

Solution:
I added a line of code to set newEntry to true when
only the container should be pre selected in the form.